Added iuse_divide_dist (also in GUI) Added PEST capability

+++ GUI/misc.h
@@ -67,7 +67,7 @@ public:
 signals:
     void display();
 //    void displayF();
-    void summary(int);
+    void summary(double);
     void setupC(int);
     void facs_update();
     void histo_update();
